Exit gcc9
Not using it for a while when deprecation message shows up:

```
Installing gcc@9
Warning: gcc@9 has been deprecated! It will be disabled on 2026-09-18.
==> Fetching downloads for: gcc@9
gcc@9: This formula either does not compile or function as expected on macOS
versions newer than Monterey due to an upstream incompatibility.
gcc@9: The x86_64 architecture is required for this software.
Error: gcc@9: Unsatisfied requirements failed this build.
Installing gcc@9 has failed!
```